Fixes non-portable way of executing pwd
Calling the `pwd` command using the absolute path `/bin/pwd` is non-portable, since the executable might be placed somewhere else depending on with Linux-distribution is being used. This commit fixes this by executing `pwd` without an absolute path, instead relying on `$PATH`. 

The reason I found this problem is that on my Linux-distribution `pwd` is not placed inside `/bin`, which resulted in this error for me:

```
bin/sstools-utils: line 8: /bin/pwd: No such file or directory
```